changed migrations to run on dedalo created db
Former-commit-id: e49bced67ce27e64ebc58f99621e387815afbbc4

+++ db/migrate/20171130090859_mv_typology_from_person_translations_to_people.rb
@@ -1,6 +1,8 @@
 class MvTypologyFromPersonTranslationsToPeople < ActiveRecord::Migration[5.0]
   def change
-    remove_column :person_translations, :typology, :string
-    add_column :people, :typology, :string
+    unless Project.name.to_sym == :mog
+      remove_column :person_translations, :typology, :string
+      add_column :people, :typology, :string
+    end
   end
 end
