+++ db/migrate/20180509134619_move_transcript_and_translation_segment_translations_text.rb
@@ -1,25 +1,27 @@
 class MoveTranscriptAndTranslationSegmentTranslationsText < ActiveRecord::Migration[5.0]
   def change
-    reversible do |dir|
-      dir.up do
-        Segment.add_translation_fields! text: :text
+    unless Project.name.to_sym == :mog
+      reversible do |dir|
+        dir.up do
+          Segment.add_translation_fields! text: :text
+        end
+
+        dir.down do
+          remove_column :segment_translations, :text
+        end
       end
 
-      dir.down do
-        remove_column :segment_translations, :text
+      Segment.find_each do |segment|
+        transcript = connection.quote segment.transcript
+        translation = connection.quote segment.attributes["translation"]
+
+        existing_locales = segment.translations.map(&:locale)
+        insert_or_update_by_sql('de', translation, segment, existing_locales)
+
+        original_lang_locales = segment.interview.language.code.split(/[\/-]/) # these are alpha3
+        original_locale = ISO_639.find(original_lang_locales[0]).alpha2
+        insert_or_update_by_sql(original_locale, transcript, segment, existing_locales)
       end
-    end
-
-    Segment.find_each do |segment|
-      transcript = connection.quote segment.transcript
-      translation = connection.quote segment.attributes["translation"]
-
-      existing_locales = segment.translations.map(&:locale)
-      insert_or_update_by_sql('de', translation, segment, existing_locales)
-
-      original_lang_locales = segment.interview.language.code.split(/[\/-]/) # these are alpha3
-      original_locale = ISO_639.find(original_lang_locales[0]).alpha2
-      insert_or_update_by_sql(original_locale, transcript, segment, existing_locales)
     end
 
     remove_column :segments, :transcript
